LayoutTests/fast/forms/interactive-validation-compat-mode-expected.txt

Test if an invalid form in a document without DOCTYPE is submitted.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS The form should be submitted.
PASS location.search.indexOf("i0=") != -1 is true
TEST COMPLETE
 
